Added queueless withdrawals for capella (#6423)
 - State was missing `next_withdrawal_index`
 - changed MaxWithdrawalsPerLayload to be int
 - had to breakout process_execution_payload logic so that I could add withdrawals into that, and the spec still does the TTD check in capella which is maybe to cover corner cases, not sure...

 fixes #6321

+++ ethereum/spec/src/main/java/tech/pegasys/teku/spec/config/SpecConfigBuilder.java
@@ -920,7 +920,7 @@ public class SpecConfigBuilder {
     private UInt64 capellaForkEpoch;
 
     private UInt64 maxBlsToExecutionChanges;
-    private UInt64 maxWithdrawalsPerPayload;
+    private int maxWithdrawalsPerPayload;
 
     private CapellaBuilder() {}
 
@@ -956,7 +956,7 @@ public class SpecConfigBuilder {
       return this;
     }
 
-    public CapellaBuilder maxWithdrawalsPerPayload(final UInt64 maxWithdrawalsPerPayload) {
+    public CapellaBuilder maxWithdrawalsPerPayload(final int maxWithdrawalsPerPayload) {
       this.maxWithdrawalsPerPayload = maxWithdrawalsPerPayload;
       return this;
     }
