+++ ethereum/spec/src/main/java/tech/pegasys/teku/spec/config/SpecConfigCapella.java
@@ -38,5 +38,5 @@ public interface SpecConfigCapella extends SpecConfigBellatrix {
 
   UInt64 getMaxBlsToExecutionChanges();
 
-  UInt64 getMaxWithdrawalsPerPayload();
+  int getMaxWithdrawalsPerPayload();
 }
