+++ ethereum/spec/src/main/java/tech/pegasys/teku/spec/datastructures/execution/versions/capella/ExecutionPayloadSchemaCapella.java
@@ -99,7 +99,7 @@ public class ExecutionPayloadSchemaCapella
         namedSchema(
             WITHDRAWALS,
             SszListSchema.create(
-                new WithdrawalSchema(), specConfig.getMaxWithdrawalsPerPayload().longValue())));
+                new WithdrawalSchema(), specConfig.getMaxWithdrawalsPerPayload())));
     this.defaultExecutionPayload = createFromBackingNode(getDefaultTree());
   }
